Fix primary index access issue

+++ business-logic/src/main/java/com/netflix/vms/transformer/SimpleTransformer.java
@@ -363,7 +363,7 @@ public class SimpleTransformer {
 
     private List<MoviePersonCharacter> getPersonCharacters(HollowPrimaryKeyIndex primaryKeyIndex, CompleteVideo completeVideo) {
         List<MoviePersonCharacter> personCharacters = new ArrayList<>();
-        int movieId = completeVideo.id.value;
+        long movieId = completeVideo.id.value;
         int matchingOrdinal = primaryKeyIndex.getMatchingOrdinal(movieId);
         if(matchingOrdinal != -1) {
             MovieCharacterPersonHollow movieCharacterPersonHollow = api.getMovieCharacterPersonHollow(matchingOrdinal);
@@ -372,7 +372,7 @@ public class SimpleTransformer {
             while(iterator.hasNext()) {
                 PersonCharacterHollow personCharacterHollow = iterator.next();
                 MoviePersonCharacter moviePersonCharacter = new MoviePersonCharacter();
-                moviePersonCharacter.movieId = movieId;
+                moviePersonCharacter.movieId = (int)movieId;
                 moviePersonCharacter.personId = (int)personCharacterHollow._getPersonId();
                 moviePersonCharacter.characterId = personCharacterHollow._getCharacterId();
                 personCharacters.add(moviePersonCharacter);
